refactored components and added absolute path resolution for components
