+++ src/pages/home/Body/index.js
@@ -1,0 +1,3 @@
+import Body from './Body';
+
+export default Body;
