+++ src/pages/home/Body/styles.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+const bodyText = { 
+  fontFamily: "Poppins",
+  fontWeight: 300,
+  marginLeft: "20%",
+  marginRight: "20%",
+  fontSize: 30,
+}
+
+export {
+  bodyText
+};
+
+// Gatsby requires a default export
+// to be a component or string
+
+const NullComponent = () => (<div></div>)
+export default NullComponent;
